[p4] remove the integer type cast in the CUDA and HIP kernels

+++ src/p4-cuda/main.cu
@@ -98,7 +98,7 @@ void postprocess (
     const float dir_rot = val - floorf(val / (period + 1e-8f)) * period;
     const float yaw = dir_rot + dir_offset + period * dir_label;
 
-    int resCount = (int)atomicAdd(object_counter, 1);
+    int resCount = atomicAdd(object_counter, 1);
     bndbox_output[0] = resCount+1;
     float *data = bndbox_output + 1 + resCount * 9;
     data[0] = box_input[box_offset];
